Adjust log exports to use JS-safe numbers

+++ src-tauri/src/main.rs
@@ -301,7 +301,7 @@ async fn mcp_get_client_tools(client_name: String) -> Result<Vec<ToolInfo>, Stri
 
 #[derive(serde::Serialize, serde::Deserialize, specta::Type)]
 struct LogsCursor {
-    ts_ms: i64,
+    ts_ms: f64,
     id: String,
 }
 
@@ -318,7 +318,10 @@ struct LogsListParams {
 #[specta::specta]
 async fn mcp_logs_list(params: LogsListParams) -> Result<Vec<logging::EventRow>, String> {
     let limit = params.limit.unwrap_or(50) as usize;
-    let after = params.after.as_ref().map(|c| (c.ts_ms, c.id.as_str()));
+    let after = params
+        .after
+        .as_ref()
+        .map(|c| (c.ts_ms as i64, c.id.as_str()));
     logging::query_events(logging::QueryParams {
         server: params.server.as_deref(),
         method: params.method.as_deref(),
@@ -330,7 +333,7 @@ async fn mcp_logs_list(params: LogsListParams) -> Result<Vec<logging::EventRow>,
 
 #[derive(serde::Serialize, serde::Deserialize, specta::Type)]
 struct LogsSinceParams {
-    since_ts_ms: i64,
+    since_ts_ms: f64,
     server: Option<String>,
     method: Option<String>,
     ok: Option<bool>,
@@ -342,7 +345,7 @@ struct LogsSinceParams {
 async fn mcp_logs_list_since(params: LogsSinceParams) -> Result<Vec<logging::EventRow>, String> {
     let limit = params.limit.unwrap_or(50) as usize;
     logging::query_events_since(
-        params.since_ts_ms,
+        params.since_ts_ms as i64,
         params.server.as_deref(),
         params.method.as_deref(),
         params.ok,
@@ -352,7 +355,7 @@ async fn mcp_logs_list_since(params: LogsSinceParams) -> Result<Vec<logging::Eve
 
 #[tauri::command]
 #[specta::specta]
-async fn mcp_logs_count(server: Option<String>) -> Result<i64, String> {
+async fn mcp_logs_count(server: Option<String>) -> Result<f64, String> {
     logging::count_events(server.as_deref())
 }
